requirements.txt
python-dotenv==0.19.2
Werkzeug==2.0.3
SQLAlchemy==1.4.31
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask==2.0.3
psycopg2-binary==2.9.3
